Working solution better performace with recursion

+++ src/copythreader.py
@@ -1,10 +1,10 @@
+import logging
 import os
+import queue
 import shutil
 import threading
-from typing import List
-import logging
-import queue
 from pathlib import Path
+from typing import List
 
 file_queue = queue.Queue()
 
@@ -18,6 +18,7 @@ class ThreadedCopy:
         self.src_dir = src_dir
         self.replica_dir = replica_dir
         self.total_files = len(files_to_copy)
+        self.copy_count = 0
         self.thread_worker_copy(files_to_copy)
 
     def copy_worker(self) -> None:
@@ -27,14 +28,13 @@ class ThreadedCopy:
             dst_path = self.replica_dir / relative_path
             dst_dir = dst_path.parent
             os.makedirs(dst_dir, exist_ok=True)  # Ensure the destination directory exists
-            shutil.copy(str(file_path), str(dst_path))
+            if dst_path.exists():
+                shutil.copy(str(file_path), str(dst_path))
+                logging.info(f"Copied {file_path} to {dst_path}")
+            else:
+                shutil.copy(str(file_path), str(dst_path))
+                logging.info(f"Created {dst_path}")
             file_queue.task_done()
-            with self.lock:
-                self.copy_count += 1
-                percent = (self.copy_count * 100) / self.total_files
-                print(str(percent) + " percent copied.")
-
-
 
     def thread_worker_copy(self, files_to_copy: List[Path]) -> None:
         for i in range(16):
